refact: signup 페이지 구조 변경
