+++ gible/src/GloablStyle.js
@@ -4,6 +4,9 @@ const GlobalStyle = createGlobalStyle`
     body {
         font-family : pretendard;
     }
-`
+    :root {
+        --main-color : #F42E3D;
+    }
+`;
 
 export default GlobalStyle;
