+++ gible/src/components/signup/step02/Step02.js
@@ -1,0 +1,14 @@
+import React from "react";
+import InformationTitle from "./InformationTitle";
+import InformationBox from "./InformationBox";
+
+const Step02 = ({ setStep }) => {
+  return (
+    <>
+      <InformationTitle />
+      <InformationBox setStep={setStep} />
+    </>
+  );
+};
+
+export default Step02;
