+++ gible/src/components/signup/step03/Step03.js
@@ -1,0 +1,16 @@
+import React from "react";
+import Welcome from "./Welcome";
+import Guide from "./Guide";
+import ToLoginBtn from "./ToLoginBtn";
+
+const Step03 = () => {
+  return (
+    <>
+      <Welcome />
+      <Guide />
+      <ToLoginBtn />
+    </>
+  );
+};
+
+export default Step03;
